test: split state.json-backed routes out of bridge-routes.test.ts
The suite crossed the 500-line cap once the SSE abort test landed; the
quick-prompts/settings/custom-engine suite moves to web-state-routes
with the shared fakes extracted to route-fakes.ts.
